+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 -----
 
 - #131 Push FHIR Observations to Tamanu using a FHIR Transaction Bundle
+- #133 Add microbiology statistic reports for contamination, pathogen, AST panel, and sample rejection
+- #132 Centralize Tamanu host and credentials in the control panel
+- #126 Disable automatic client creation on Tamanu sync by default
 - #124 Add result variables and improve result formatting to Tupaia export file
 - #119 Fix empty pdf when notifying DiagnosticReport to Tamanu
 - #116 Compatibility with core#2831 (Migrate ARReport to Dexterity)